+++ src/python-fastui/fastui/components/__init__.py
@@ -12,6 +12,7 @@ from pydantic_core import core_schema as _core_schema
 from .. import class_name as _class_name
 from .. import events
 from .. import types as _types
+from ..base import BaseModel
 from .display import Details, Display
 from .forms import (
     Form,
@@ -23,7 +24,6 @@ from .forms import (
     FormFieldSelectSearch,
     ModelForm,
 )
-from .shared_base import BaseModel
 from .tables import Pagination, Table
 
 __all__ = (
